Use PasteContext for annotations (in a nested way)
Now annotations inside PasteContext have their own “original version of
the annotation”, “latest version of the annotation”, etc.

This commit doesn't change any behavior, but makes it more obvious that
the current behavior is wrong (e.g. we use the pcOriginal field of each
annotation, but we should be using pcLatest just as we do for pastes themselves).