Don't pass the full list of annotations
Previously we were passing a list of annotations to viewPaste (and
pasteDetails, and pasteNav), but it's not needed – the only place where
annotations are actually required is pasteNav, and what pasteNav wants
to know is:

  * the original paste that everyone else annotates
  * the previous annotation

Both these things can be easily generated in viewAnnotations by doing
some zipping (instead of having to do a search with 'find' in pasteNav).

This commit doesn't change any behavior.